Refactor code to ES6 modules
