+++ index.html
@@ -7,7 +7,7 @@
     <title>Awesome Books</title>
 
     <!-- Custom Stylesheets -->
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="index.css" />
   </head>
   <body>
     <!-- Main section starts here -->
@@ -30,9 +30,6 @@
         <section class="books">
           <h1 class="title">All awesome books</h1>
           <ul class="books">
-            <li class="books">
-              Book Name <button class="delete">Delete</button>
-            </li>
           </ul>
         </section>
         <section class="add-book dsp-none">
@@ -61,6 +58,6 @@
     <!-- Main section ends here -->
 
     <!-- Custom JS -->
-    <script src="main.js"></script>
+    <script src="index.js" type="module"></script>
   </body>
 </html>
