+++ modules/dom-elements.js
@@ -1,0 +1,8 @@
+export const bookList = document.querySelector('ul.books');
+export const addForm = document.querySelector('form.add-book');
+export const storageDetails = localStorage.getItem('books');
+export const timeText = document.querySelector('.time');
+export const navbar = document.querySelector('ul.navbar');
+export const bookSection = document.querySelector('section.books');
+export const addBookSection = document.querySelector('section.add-book');
+export const contactSection = document.querySelector('section.contact');
